docker entry script tests

+++ docker_entry_script.sh
@@ -3,17 +3,21 @@
 echo "Generating config..."
 python /tryton/trytond-scripts/generate_config.py
 
-echo "Generating password file..."
-echo "$TRYTONPASS" > $TRYTONPASSFILE
+if [ $# -gt 0 ]; then
+    echo "First run."
 
-echo "Initializing the DB..."
-/usr/local/bin/trytond -c $TRYTOND_CONFIG -d $DATABASE_NAME -v --all
+    echo "Generating password file..."
+    echo "$TRYTONPASS" > $TRYTONPASSFILE
 
-echo "Removing password file..."
-rm $TRYTONPASSFILE
+    echo "Initializing the DB..."
+    /usr/local/bin/trytond -c $TRYTOND_CONFIG -d $DATABASE_NAME -v --all
 
-echo "Install ALL the modules..."
-python /tryton/trytond-scripts/install_modules.py
+    echo "Removing password file..."
+    rm $TRYTONPASSFILE
+
+    echo "Install ALL the modules..."
+    python /tryton/trytond-scripts/install_modules.py
+fi
 
 echo "Launching Trytond Server..."
 exec /usr/local/bin/trytond -c $TRYTOND_CONFIG -d $DATABASE_NAME -v
